+++ eoreader/__meta__.py
@@ -17,4 +17,4 @@
 **EOReader** library
 """
 
-__version__ = "0.21.9"
+__version__ = "0.21.10"