+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 0.21.10 (2024-mm-dd)
 
 - ENH: Allow the user to change the default resampling per product and free the `resolution` keyword in `load` and `stack` ([#103](https://github.com/sertit/eoreader/discussions/103))
+- ENH: For Sentinel-2 L2A products, use the `MSK_CLDPRB_20m` for computing clouds, as they are more precise than legacy L1C `MSK_CLOUDS_B00` masks. ([#174](https://github.com/sertit/eoreader/issues/174))
 - FIX: Fix Sentinel-2 Theia footprints when the nodata area is wider than the data area ([#201](https://github.com/sertit/eoreader/issues/201))
 - FIX: Pop `driver` keyword in stack function to only use it for writing, allowing people to drive stack as COGs ([#181](https://github.com/sertit/eoreader/issues/181), [#202](https://github.com/sertit/eoreader/discussions/202))
 - FIX: Simplify L7 footprint ([#198](https://github.com/sertit/eoreader/issues/198))